Add QoiImageFormat
