+++ image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/qoi/Header.scala
@@ -1,0 +1,9 @@
+package eu.joaocosta.minart.graphics.image.qoi
+
+final case class Header(
+    magic: String,
+    width: Long,
+    height: Long,
+    channels: Byte,
+    colorspace: Byte
+)
